Refactor GetIt usage
