+++ pubspec.yaml
@@ -36,6 +36,7 @@ dev_dependencies:
     sdk: flutter
   flutter_lints: ^4.0.0
   json_serializable: ^6.8.0
+  mocktail: ^1.0.4
   retrofit_generator: ^8.1.0
 
 flutter:
